Make title link better.
To make the whole coloured block feel linkish, I've moved the padding
from the outer div to the link element itself, and made that a block so
that it can have an effective background colour. This means that changing
the background colour of the link on hover changes the whole thing.